Refactor rewards and shop features
