+++ App.js
@@ -5,15 +5,11 @@ import {
   Text,
   ScrollView,
   TouchableOpacity,
-  TouchableWithoutFeedback,
   StyleSheet,
   SafeAreaView,
   StatusBar,
   Platform,
-  Modal,
   Alert,
-  Share,
-  PanResponder,
   Keyboard,
   ActivityIndicator,
 } from 'react-native';
@@ -31,10 +27,6 @@ import {
   focusCost,
   computeRewards,
   GAME_EFFECTS,
-  REAL_REWARDS,
-  PREMIUM_REWARDS,
-  buyEffect,
-  redeemReward,
 } from './src/features/progression';
 import { STATUSES } from './src/features/applications';
 import { QUESTS, composeQuestClaimKey, CLAIM_KEY_SEPARATOR } from './src/features/quests';
@@ -68,6 +60,20 @@ import { IconButton } from './src/components/layout/IconButton';
 import { ProgressBar } from './src/components/layout/ProgressBar';
 import { EffectTimerRing } from './src/components/feedback/EffectTimerRing';
 import { ToastBanner } from './src/components/feedback/ToastBanner';
+import {
+  useRewardsInventory,
+  RewardSummary,
+  ChestGrid,
+  createLevelUpChest,
+} from './src/features/rewards';
+import {
+  ShopScreen,
+  includeSprayEffect,
+  createSprayEffectDetails,
+  isSprayDebuffActive,
+  shouldClearSprayDebuff,
+  SPRAY_DEBUFF_DURATION_MS,
+} from './src/features/shop';
 
 const APP_VERSION = appConfig?.expo?.version ?? '1.0.0';
 
@@ -133,20 +139,6 @@ const BOTTOM_TABS = [
   { key: 'Shop', label: 'Shop', icon: 'shopping-outline' },
 ];
 
-const SHOP_MAIN_TABS = [
-  { key: 'active', label: 'Active', icon: 'auto-fix' },
-  { key: 'catalogue', label: 'Catalogue', icon: 'view-grid-outline' },
-];
-
-const SHOP_CATALOG_TABS = [
-  { key: 'effects', label: 'Boosts', icon: 'flash-outline' },
-  { key: 'rewards', label: 'Treats', icon: 'cupcake' },
-  { key: 'premium', label: 'Premium', icon: 'crown-outline' },
-];
-
-const SPRAY_DEBUFF_DURATION_MS = 6 * 60 * 60 * 1000;
-const SPRAY_DEBUFF_DURATION_SECONDS = SPRAY_DEBUFF_DURATION_MS / 1000;
-
 const STATUS_META = {
   Applied: { icon: 'file-document-outline', tint: 'sky' },
   'Applied with referral': { icon: 'account-star-outline', tint: 'sky' },
@@ -159,1380 +151,9 @@ const LOG_STATUS_OPTIONS = STATUSES.filter(
   (status) => status.key === 'Applied' || status.key === 'Applied with referral',
 );
 
-const costFor = (item) => Math.round(item.minutes * (item.pleasure ?? 1));
-
 const STORAGE_KEY = 'jobless::state';
 const STORAGE_VERSION = 1;
 
-const createChestArt = ({
-  baseGradient,
-  lidGradient,
-  outline,
-  strapColor,
-  strapHighlight,
-  latchColor,
-  latchOutline,
-  accentLine,
-  baseDividers,
-}) => {
-  return ({ size = 56 }) => {
-    const baseWidth = size * 0.82;
-    const baseHeight = size * 0.48;
-    const lidHeight = size * 0.3;
-    const strapWidth = size * 0.18;
-    const strapHeight = size * 0.56;
-    const latchWidth = size * 0.24;
-    const latchHeight = size * 0.24;
-    const radius = size * 0.18;
-    const outlineWidth = Math.max(1, Math.round(size * 0.035));
-    const baseLeft = (size - baseWidth) / 2;
-    const lidTop = size * 0.12;
-    const baseBottom = size * 0.16;
-
-    return (
-      <View
-        style={{
-          width: size,
-          height: size,
-          alignItems: 'center',
-          justifyContent: 'center',
-        }}
-      >
-        <View
-          style={{
-            position: 'absolute',
-            bottom: baseBottom - baseHeight * 0.1,
-            width: baseWidth * 1.04,
-            height: baseHeight * 0.22,
-            backgroundColor: outline,
-            opacity: 0.18,
-            borderRadius: baseHeight * 0.3,
-            transform: [{ scaleY: 0.8 }],
-          }}
-        />
-        <LinearGradient
-          colors={lidGradient}
-          start={{ x: 0, y: 0 }}
-          end={{ x: 1, y: 1 }}
-          style={{
-            position: 'absolute',
-            top: lidTop,
-            left: baseLeft,
-            width: baseWidth,
-            height: lidHeight,
-            borderTopLeftRadius: radius,
-            borderTopRightRadius: radius,
-            borderBottomLeftRadius: radius * 0.6,
-            borderBottomRightRadius: radius * 0.6,
-            borderWidth: outlineWidth,
-            borderColor: outline,
-          }}
-        />
-        <LinearGradient
-          colors={baseGradient}
-          start={{ x: 0, y: 0 }}
-          end={{ x: 1, y: 1 }}
-          style={{
-            position: 'absolute',
-            bottom: baseBottom,
-            left: baseLeft,
-            width: baseWidth,
-            height: baseHeight,
-            borderRadius: radius,
-            borderWidth: outlineWidth,
-            borderColor: outline,
-          }}
-        />
-        {(baseDividers || []).map((offset, index) => (
-          <View
-            key={`divider-${index}`}
-            style={{
-              position: 'absolute',
-              bottom: baseBottom + outlineWidth,
-              left: baseLeft + baseWidth * offset - outlineWidth / 2,
-              width: outlineWidth,
-              height: baseHeight - outlineWidth * 2,
-              backgroundColor: outline,
-              opacity: 0.22,
-              borderRadius: outlineWidth,
-            }}
-          />
-        ))}
-        <View
-          style={{
-            position: 'absolute',
-            top: size * 0.1,
-            left: size / 2 - strapWidth / 2,
-            width: strapWidth,
-            height: strapHeight,
-            borderRadius: strapWidth / 2,
-            backgroundColor: strapColor,
-            borderWidth: outlineWidth,
-            borderColor: outline,
-            justifyContent: 'center',
-          }}
-        >
-          {strapHighlight ? (
-            <View
-              style={{
-                alignSelf: 'center',
-                width: strapWidth * 0.36,
-                height: strapHeight - outlineWidth * 3,
-                borderRadius: strapWidth / 2,
-                backgroundColor: strapHighlight,
-                opacity: 0.85,
-              }}
-            />
-          ) : null}
-        </View>
-        {accentLine ? (
-          <View
-            style={{
-              position: 'absolute',
-              bottom: size * (accentLine.offset ?? 0.26),
-              width: baseWidth * (accentLine.width ?? 0.56),
-              left:
-                size / 2 - (baseWidth * (accentLine.width ?? 0.56)) / 2,
-              height: Math.max(1, size * (accentLine.thickness ?? 0.05)),
-              borderRadius: size * (accentLine.radius ?? 0.04),
-              backgroundColor: accentLine.color,
-              opacity: accentLine.opacity ?? 0.35,
-            }}
-          />
-        ) : null}
-        <View
-          style={{
-            position: 'absolute',
-            bottom: size * 0.2,
-            width: latchWidth,
-            height: latchHeight,
-            borderRadius: latchWidth * 0.3,
-            backgroundColor: latchColor,
-            borderWidth: outlineWidth,
-            borderColor: latchOutline || outline,
-            justifyContent: 'center',
-            alignItems: 'center',
-          }}
-        >
-          <View
-            style={{
-              width: latchWidth * 0.34,
-              height: latchHeight * 0.5,
-              borderRadius: latchWidth * 0.18,
-              backgroundColor: 'rgba(255, 255, 255, 0.6)',
-            }}
-          />
-        </View>
-      </View>
-    );
-  };
-};
-
-const CommonChest = createChestArt({
-  baseGradient: ['#D7B48A', '#B28755'],
-  lidGradient: ['#E4C59B', '#C89B69'],
-  outline: 'rgba(78, 54, 33, 0.65)',
-  strapColor: '#8D5A2B',
-  strapHighlight: 'rgba(247, 230, 200, 0.55)',
-  latchColor: '#F7E6C8',
-  baseDividers: [0.32, 0.68],
-});
-
-const RareChest = createChestArt({
-  baseGradient: ['#8BB7D8', '#4D7BA6'],
-  lidGradient: ['#A7CCE4', '#5C8FB6'],
-  outline: 'rgba(31, 58, 82, 0.75)',
-  strapColor: '#1F4D6F',
-  strapHighlight: 'rgba(210, 230, 244, 0.55)',
-  latchColor: '#E8F1F7',
-  latchOutline: 'rgba(31, 58, 82, 0.75)',
-  baseDividers: [0.28, 0.72],
-  accentLine: {
-    color: 'rgba(31, 58, 82, 0.72)',
-    opacity: 0.4,
-    width: 0.6,
-    thickness: 0.045,
-    offset: 0.27,
-  },
-});
-
-const EpicChest = createChestArt({
-  baseGradient: ['#C2A5E6', '#8A6BBF'],
-  lidGradient: ['#D7C3F0', '#9A7BD0'],
-  outline: 'rgba(59, 44, 84, 0.78)',
-  strapColor: '#4E3A75',
-  strapHighlight: 'rgba(238, 225, 255, 0.45)',
-  latchColor: '#F4ECFF',
-  latchOutline: 'rgba(59, 44, 84, 0.78)',
-  baseDividers: [0.26, 0.74],
-  accentLine: {
-    color: 'rgba(59, 44, 84, 0.6)',
-    opacity: 0.55,
-    width: 0.32,
-    thickness: 0.04,
-    offset: 0.29,
-  },
-});
-
-const LegendaryChest = createChestArt({
-  baseGradient: ['#F5D48D', '#D4A440'],
-  lidGradient: ['#FFE6A8', '#E7BE55'],
-  outline: 'rgba(124, 90, 24, 0.78)',
-  strapColor: '#8A611E',
-  strapHighlight: 'rgba(255, 248, 210, 0.65)',
-  latchColor: '#FFF6D9',
-  latchOutline: 'rgba(124, 90, 24, 0.78)',
-  baseDividers: [0.26, 0.74],
-  accentLine: {
-    color: 'rgba(124, 90, 24, 0.6)',
-    opacity: 0.5,
-    width: 0.72,
-    thickness: 0.05,
-    offset: 0.23,
-  },
-});
-
-const CHEST_ART = {
-  Common: CommonChest,
-  Rare: RareChest,
-  Epic: EpicChest,
-  Legendary: LegendaryChest,
-};
-
-const RARITIES = [
-  { key: 'Common', weight: 0.52, level100Weight: 0.3, gold: [8, 24] },
-  { key: 'Rare', weight: 0.3, level100Weight: 0.35, gold: [16, 36] },
-  { key: 'Epic', weight: 0.14, level100Weight: 0.25, gold: [28, 56] },
-  { key: 'Legendary', weight: 0.04, level100Weight: 0.1, gold: [48, 96] },
-];
-
-const RARITY_DETAILS = {
-  Common: {
-    headline: 'Reliable stash',
-    helper: 'Solid boosts to keep momentum steady.',
-  },
-  Rare: {
-    headline: 'Shiny find',
-    helper: 'Elevated rewards with a spark of luck.',
-  },
-  Epic: {
-    headline: 'Elite haul',
-    helper: 'High-tier loot for big progress leaps.',
-  },
-  Legendary: {
-    headline: 'Mythic treasure',
-    helper: 'Top-shelf rewards reserved for heroes.',
-  },
-};
-
-const RARITY_LOOKUP = RARITIES.reduce((acc, entry) => {
-  acc[entry.key.toLowerCase()] = entry;
-  return acc;
-}, {});
-
-let chestIdCounter = 0;
-
-const createChestId = () => {
-  chestIdCounter += 1;
-  return `chest-${Date.now()}-${chestIdCounter}`;
-};
-
-const createChestFromRarity = (rarityInput) => {
-  if (!rarityInput) {
-    return null;
-  }
-  let key = null;
-  if (typeof rarityInput === 'string') {
-    key = rarityInput.trim();
-  } else if (typeof rarityInput === 'object') {
-    key = rarityInput.key || rarityInput.rarity || null;
-  }
-  if (!key) {
-    return null;
-  }
-  const match = RARITY_LOOKUP[key.toLowerCase()];
-  if (!match) {
-    return null;
-  }
-  return {
-    id: createChestId(),
-    rarity: match.key,
-    gold: Array.isArray(match.gold) ? [...match.gold] : match.gold,
-  };
-};
-
-const CHEST_LEVEL_CAP = 100;
-
-const clampDropLevel = (level) => {
-  if (!Number.isFinite(level)) {
-    return 1;
-  }
-  return Math.min(Math.max(Math.floor(level), 1), CHEST_LEVEL_CAP);
-};
-
-const getChestWeightsForLevel = (level) => {
-  const clampedLevel = clampDropLevel(level);
-  const progress = CHEST_LEVEL_CAP <= 1 ? 1 : (clampedLevel - 1) / (CHEST_LEVEL_CAP - 1);
-  const weights = RARITIES.map((rarity) => {
-    const start = rarity.weight ?? 0;
-    const end = rarity.level100Weight ?? start;
-    const value = start + (end - start) * progress;
-    return { key: rarity.key, weight: value };
-  });
-  const total = weights.reduce((sum, entry) => sum + entry.weight, 0);
-  if (total <= 0) {
-    return weights.map((entry) => ({ ...entry, weight: 0 }));
-  }
-  return weights.map((entry) => ({ ...entry, weight: entry.weight / total }));
-};
-
-const pickChestRarityForLevel = (level) => {
-  const weights = getChestWeightsForLevel(level);
-  if (!weights.length) {
-    return null;
-  }
-  const roll = Math.random();
-  let cumulative = 0;
-  for (const entry of weights) {
-    cumulative += entry.weight;
-    if (roll < cumulative) {
-      return entry.key;
-    }
-  }
-  return weights[weights.length - 1].key;
-};
-
-const createLevelUpChest = (level) => {
-  const rarityKey = pickChestRarityForLevel(level);
-  return rarityKey ? createChestFromRarity(rarityKey) : null;
-};
-
-const rand = ([min, max]) => Math.floor(Math.random() * (max - min + 1)) + min;
-
-const computePotential = (list = []) => {
-  if (!list.length) {
-    return null;
-  }
-  const goldMin = list.reduce(
-    (acc, chest) => acc + (Array.isArray(chest.gold) ? chest.gold[0] ?? 0 : 0),
-    0,
-  );
-  const goldMax = list.reduce(
-    (acc, chest) => acc + (Array.isArray(chest.gold) ? chest.gold[1] ?? chest.gold[0] ?? 0 : 0),
-    0,
-  );
-  return [goldMin, goldMax];
-};
-
-const formatRange = (range) => {
-  if (!range) {
-    return '0';
-  }
-  const [min, max] = range;
-  return min === max ? `${min}` : `${min} – ${max}`;
-};
-
-const GoldSlider = ({ value, max, colors, eff, onChange }) => {
-  const [trackWidth, setTrackWidth] = useState(0);
-  const clampedMax = Math.max(0, max);
-  const clampedValue = Math.max(0, Math.min(value, clampedMax));
-  const percentage = clampedMax > 0 ? Math.min(100, (clampedValue / clampedMax) * 100) : 0;
-
-  const handleGesture = useCallback(
-    (evt) => {
-      if (!trackWidth || clampedMax <= 0) {
-        return;
-      }
-      const x = evt.nativeEvent.locationX;
-      const ratio = Math.max(0, Math.min(1, x / trackWidth));
-      const next = Math.round(clampedMax * ratio);
-      onChange(next);
-    },
-    [trackWidth, clampedMax, onChange],
-  );
-
-  const panResponder = useMemo(
-    () =>
-      PanResponder.create({
-        onStartShouldSetPanResponder: () => true,
-        onMoveShouldSetPanResponder: () => true,
-        onPanResponderGrant: handleGesture,
-        onPanResponderMove: handleGesture,
-        onPanResponderRelease: handleGesture,
-      }),
-    [handleGesture],
-  );
-
-  const thumbSize = 22;
-  const thumbLeft = trackWidth ? (percentage / 100) * trackWidth : 0;
-
-  return (
-    <View
-      style={[
-        styles.shopSliderTrack,
-        {
-          borderColor: colors.surfaceBorder,
-          backgroundColor: hexToRgba(colors.text, eff === 'light' ? 0.12 : 0.28),
-        },
-      ]}
-      onLayout={(event) => setTrackWidth(event.nativeEvent.layout.width)}
-      {...panResponder.panHandlers}
-    >
-      <LinearGradient
-        colors={[colors.sky, colors.emerald]}
-        start={{ x: 0, y: 0 }}
-        end={{ x: 1, y: 0 }}
-        style={[styles.shopSliderFill, { width: `${percentage}%` }]}
-      />
-      <View
-        style={[
-          styles.shopSliderThumb,
-          {
-            left: Math.max(0, Math.min(trackWidth - thumbSize, thumbLeft - thumbSize / 2)),
-            width: thumbSize,
-            height: thumbSize,
-            borderRadius: thumbSize / 2,
-            backgroundColor: colors.surface,
-            borderColor: colors.surfaceBorder,
-          },
-        ]}
-      />
-    </View>
-  );
-};
-
-const ShopScreen = ({
-  colors,
-  eff,
-  gold,
-  setGold,
-  effects,
-  setEffects,
-  premiumProgress,
-  setPremiumProgress,
-  mainTab,
-  setMainTab,
-  categoryTab,
-  setCategoryTab,
-}) => {
-  const [savingItem, setSavingItem] = useState(null);
-  const [saveAmount, setSaveAmount] = useState(0);
-  const [confirmReward, setConfirmReward] = useState(null);
-  const [redeemed, setRedeemed] = useState(null);
-  const [now, setNow] = useState(() => Date.now());
-  const hasTimers = useMemo(() => effects.some((effect) => effect.expiresAt), [effects]);
-
-  useEffect(() => {
-    if (!hasTimers) {
-      return;
-    }
-    setNow(Date.now());
-    const interval = setInterval(() => {
-      setNow(Date.now());
-    }, 1000);
-    return () => clearInterval(interval);
-  }, [hasTimers]);
-
-  const sortedRewards = useMemo(
-    () => REAL_REWARDS.slice().sort((a, b) => costFor(a) - costFor(b)),
-    [],
-  );
-
-  const cardShadow = eff === 'light' ? styles.shopCardShadowLight : styles.shopCardShadowDark;
-  const filledSurface = useMemo(
-    () => (eff === 'light' ? ensureOpaque(colors.surface) : colors.surface),
-    [eff, colors.surface],
-  );
-
-  const handleBuyEffect = useCallback(
-    (item) => {
-      buyEffect(item, gold, setGold, effects, setEffects);
-    },
-    [gold, setGold, effects, setEffects],
-  );
-
-  const handleRedeemReward = useCallback(
-    (item) => {
-      redeemReward(item, gold, setGold, setRedeemed);
-    },
-    [gold, setGold],
-  );
-
-  const handlePremiumAction = useCallback(
-    (item) => {
-      const cost = costFor(item);
-      const progress = premiumProgress[item.id] || 0;
-      if (progress >= cost) {
-        setConfirmReward({ ...item, premium: true });
-      } else {
-        setSavingItem(item);
-        setSaveAmount(Math.min(gold, cost - progress));
-      }
-    },
-    [premiumProgress, gold],
-  );
-
-  const closeSavingModal = useCallback(() => {
-    setSavingItem(null);
-    setSaveAmount(0);
-  }, []);
-
-  const confirmSave = useCallback(() => {
-    if (!savingItem) {
-      return;
-    }
-    const cost = costFor(savingItem);
-    const progress = premiumProgress[savingItem.id] || 0;
-    const maxSavable = Math.max(0, Math.min(gold, cost - progress));
-    const amount = Math.max(0, Math.min(saveAmount, maxSavable));
-    if (amount > 0) {
-      setGold((g) => g - amount);
-      setPremiumProgress((prev) => ({
-        ...prev,
-        [savingItem.id]: Math.min(cost, progress + amount),
-      }));
-    }
-    setSavingItem(null);
-    setSaveAmount(0);
-  }, [savingItem, premiumProgress, gold, saveAmount, setGold, setPremiumProgress]);
-
-  const confirmCost = confirmReward ? costFor(confirmReward) : 0;
-  const ConfirmIcon = confirmReward?.premium ? 'crown' : 'gift-outline';
-
-  const savingProgress = savingItem ? premiumProgress[savingItem.id] || 0 : 0;
-  const savingCap = savingItem ? Math.max(costFor(savingItem) - savingProgress, 0) : 0;
-  const maxSavableNow = savingItem ? Math.max(0, Math.min(gold, savingCap)) : 0;
-  const clampedSaveAmount = Math.max(0, Math.min(saveAmount, maxSavableNow));
-
-  const handleConfirmAction = useCallback(() => {
-    if (!confirmReward) {
-      return;
-    }
-    if (confirmReward.premium) {
-      setRedeemed(confirmReward);
-      setPremiumProgress((prev) => ({
-        ...prev,
-        [confirmReward.id]: 0,
-      }));
-    } else {
-      handleRedeemReward(confirmReward);
-    }
-    setConfirmReward(null);
-  }, [confirmReward, handleRedeemReward, setPremiumProgress]);
-
-  const formatDuration = useCallback((duration) => {
-    if (!duration) {
-      return 'Instant';
-    }
-    if (duration >= 3600) {
-      return `${Math.round(duration / 3600)} hr`;
-    }
-    return `${Math.round(duration / 60)} min`;
-  }, []);
-
-  const overlayBackground = eff === 'light' ? 'rgba(0,0,0,0.25)' : 'rgba(0,0,0,0.55)';
-
-  return (
-    <>
-      <ScrollView
-        style={styles.content}
-        contentContainerStyle={styles.scrollContent}
-        showsVerticalScrollIndicator={false}
-      >
-        <View
-          style={[styles.shopContainer, { borderColor: colors.surfaceBorder, backgroundColor: colors.surface }]}
-        >
-          <LinearGradient
-            colors={[hexToRgba(colors.rose, 0.28), 'transparent']}
-            start={{ x: 0, y: 0 }}
-            end={{ x: 1, y: 1 }}
-            style={styles.shopGlowTopLeft}
-            pointerEvents="none"
-          />
-          <LinearGradient
-            colors={[hexToRgba(colors.sky, 0.28), 'transparent']}
-            start={{ x: 1, y: 0 }}
-            end={{ x: 0, y: 1 }}
-            style={styles.shopGlowTopRight}
-            pointerEvents="none"
-          />
-          <LinearGradient
-            colors={[hexToRgba(colors.emerald, 0.24), 'transparent']}
-            start={{ x: 1, y: 1 }}
-            end={{ x: 0, y: 0 }}
-            style={styles.shopGlowBottomRight}
-            pointerEvents="none"
-          />
-          <LinearGradient
-            colors={[hexToRgba(colors.lilac, 0.26), 'transparent']}
-            start={{ x: 0, y: 1 }}
-            end={{ x: 1, y: 0 }}
-            style={styles.shopGlowBottomLeft}
-            pointerEvents="none"
-          />
-          <View style={styles.shopContent}>
-            <View
-              style={[
-                styles.shopMainTabs,
-                { backgroundColor: colors.surface, borderColor: colors.surfaceBorder },
-              ]}
-            >
-              {SHOP_MAIN_TABS.map((tab) => {
-                const active = mainTab === tab.key;
-                return (
-                  <TouchableOpacity
-                    key={tab.key}
-                    onPress={() => setMainTab(tab.key)}
-                    activeOpacity={0.9}
-                    style={[styles.shopMainTabButton, { borderColor: colors.surfaceBorder }]}
-                  >
-                    {active ? (
-                      <LinearGradient
-                        colors={[colors.sky, colors.emerald]}
-                        start={{ x: 0, y: 0 }}
-                        end={{ x: 1, y: 0 }}
-                        style={StyleSheet.absoluteFillObject}
-                      />
-                    ) : null}
-                    <View style={styles.shopMainTabContent}>
-                      <MaterialCommunityIcons
-                        name={tab.icon}
-                        size={16}
-                        color={active ? '#0f172a' : colors.text}
-                        style={styles.shopMainTabIcon}
-                      />
-                      <Text
-                        style={[
-                          styles.shopMainTabLabel,
-                          { color: active ? '#0f172a' : colors.text },
-                        ]}
-                      >
-                        {tab.label}
-                      </Text>
-                    </View>
-                  </TouchableOpacity>
-                );
-              })}
-            </View>
-
-            {mainTab === 'active' ? (
-              <View style={styles.shopSection}>
-                <View style={styles.shopSectionHeader}>
-                  <Text style={[styles.shopSectionEyebrow, { color: hexToRgba(colors.text, 0.55) }]}>Active effects</Text>
-                  <Text style={[styles.shopSectionTitle, { color: colors.text }]}>Track your current boosts and penalties in real time.</Text>
-                </View>
-                {effects.length === 0 ? (
-                  <View
-                    style={[
-                      styles.shopEmptyCard,
-                      { backgroundColor: filledSurface, borderColor: colors.surfaceBorder },
-                      cardShadow,
-                    ]}
-                  >
-                    <View style={styles.shopEmptyHeader}>
-                      <MaterialCommunityIcons name="auto-fix" size={18} color={colors.sky} />
-                      <Text style={[styles.shopEmptyTitle, { color: colors.text }]}>No active effects yet</Text>
-                    </View>
-                    <Text style={[styles.shopEmptyDescription, { color: hexToRgba(colors.text, 0.65) }]}>Activate a boost to double down on XP or gold. Your effects — including any penalties — will appear here with live timers once purchased.</Text>
-                    <TouchableOpacity
-                      onPress={() => {
-                        setMainTab('catalogue');
-                        setCategoryTab('effects');
-                      }}
-                      activeOpacity={0.9}
-                      style={[styles.shopBrowseButton, { borderColor: colors.surfaceBorder }]}
-                    >
-                      <LinearGradient
-                        colors={[colors.sky, colors.emerald]}
-                        start={{ x: 0, y: 0 }}
-                        end={{ x: 1, y: 0 }}
-                        style={StyleSheet.absoluteFillObject}
-                      />
-                      <Text style={styles.shopBrowseButtonText}>Browse boosts</Text>
-                    </TouchableOpacity>
-                  </View>
-                ) : (
-                  <View style={styles.shopActiveList}>
-                    {effects.map((effect) => {
-                      const remaining = effect.expiresAt
-                        ? Math.max(0, Math.floor((effect.expiresAt - now) / 1000))
-                        : null;
-                      const progress =
-                        effect.duration && remaining != null
-                          ? Math.max(0, Math.min(1, remaining / effect.duration))
-                          : null;
-                      const isDebuff = effect.type === 'debuff';
-                      const cardBackgroundColor = filledSurface;
-                      const cardBorderColor = colors.surfaceBorder;
-                      const iconShellStyle = {
-                        borderColor: colors.surfaceBorder,
-                        backgroundColor: colors.surface,
-                      };
-                      const iconGradient = [
-                        hexToRgba(colors.surface, 0.98),
-                        hexToRgba(colors.chipBg, 0.85),
-                      ];
-                      const iconColor = '#0f172a';
-                      const timerColor = hexToRgba(colors.text, 0.6);
-                      const passiveColor = hexToRgba(colors.text, 0.6);
-                      const descriptionColor = hexToRgba(colors.text, 0.65);
-                      const nameColor = colors.text;
-
-                      return (
-                        <View
-                          key={effect.id}
-                          style={[
-                            styles.shopActiveCard,
-                            { backgroundColor: cardBackgroundColor, borderColor: cardBorderColor },
-                            cardShadow,
-                          ]}
-                        >
-                          <EffectTimerRing
-                            progress={progress}
-                            colors={colors}
-                            eff={eff}
-                            gradientColors={isDebuff ? [colors.rose, colors.lilac] : undefined}
-                          >
-                            <View style={[styles.shopActiveIconShell, iconShellStyle]}>
-                              <LinearGradient
-                                colors={iconGradient}
-                                start={{ x: 0, y: 0 }}
-                                end={{ x: 1, y: 1 }}
-                                style={styles.shopActiveIconInner}
-                              >
-                                <MaterialCommunityIcons
-                                  name={effect.icon || 'flash-outline'}
-                                  size={22}
-                                  color={iconColor}
-                                />
-                              </LinearGradient>
-                            </View>
-                          </EffectTimerRing>
-                          <View style={styles.shopActiveInfo}>
-                            <View style={styles.shopActiveNameRow}>
-                              <Text style={[styles.shopActiveName, { color: nameColor }]}>{effect.name}</Text>
-                            </View>
-                            {remaining != null && effect.duration ? (
-                              <View style={styles.shopActiveTimerRow}>
-                                <MaterialCommunityIcons
-                                  name="clock-time-four-outline"
-                                  size={14}
-                                  color={timerColor}
-                                />
-                                <Text style={[styles.shopActiveTimerText, { color: timerColor }]}>
-                                  {formatTime(remaining)}
-                                </Text>
-                              </View>
-                            ) : (
-                              <Text style={[styles.shopActivePassive, { color: passiveColor }]}>
-                                {isDebuff ? 'Active penalty' : 'Passive boost'}
-                              </Text>
-                            )}
-                            <Text
-                              style={[styles.shopActiveDescription, { color: descriptionColor }]}
-                              numberOfLines={2}
-                            >
-                              {effect.description}
-                            </Text>
-                          </View>
-                        </View>
-                      );
-                    })}
-                  </View>
-                )}
-              </View>
-            ) : (
-              <View style={styles.shopSection}>
-                <View style={styles.shopSectionHeader}>
-                  <Text style={[styles.shopSectionEyebrow, { color: hexToRgba(colors.text, 0.55) }]}>Catalogue</Text>
-                  <Text style={[styles.shopSectionTitle, { color: colors.text }]}>Choose how you want to level the journey today.</Text>
-                </View>
-                <View
-                  style={[
-                    styles.shopSecondaryTabs,
-                    { backgroundColor: colors.surface, borderColor: colors.surfaceBorder },
-                  ]}
-                >
-                  {SHOP_CATALOG_TABS.map((tab) => {
-                    const active = categoryTab === tab.key;
-                    return (
-                      <TouchableOpacity
-                        key={tab.key}
-                        onPress={() => setCategoryTab(tab.key)}
-                        activeOpacity={0.9}
-                        style={[styles.shopSecondaryTabButton, { borderColor: colors.surfaceBorder }]}
-                      >
-                        {active ? (
-                          <LinearGradient
-                            colors={[colors.sky, colors.emerald]}
-                            start={{ x: 0, y: 0 }}
-                            end={{ x: 1, y: 0 }}
-                            style={StyleSheet.absoluteFillObject}
-                          />
-                        ) : null}
-                        <View style={styles.shopSecondaryTabContent}>
-                          <MaterialCommunityIcons
-                            name={tab.icon}
-                            size={15}
-                            color={active ? '#0f172a' : colors.text}
-                            style={styles.shopSecondaryTabIcon}
-                          />
-                          <Text
-                            style={[
-                              styles.shopSecondaryTabLabel,
-                              { color: active ? '#0f172a' : colors.text },
-                            ]}
-                          >
-                            {tab.label}
-                          </Text>
-                        </View>
-                      </TouchableOpacity>
-                    );
-                  })}
-                </View>
-
-                {categoryTab === 'effects' && (
-                  <View style={styles.shopCardList}>
-                    {GAME_EFFECTS.slice()
-                      .sort((a, b) => a.cost - b.cost)
-                      .map((item) => {
-                        const active = effects.some((fx) => fx.id === item.id);
-                        const durationLabel = formatDuration(item.duration);
-                        return (
-                          <View
-                            key={item.id}
-                            style={[
-                              styles.shopCard,
-                              { backgroundColor: filledSurface, borderColor: colors.surfaceBorder },
-                              cardShadow,
-                            ]}
-                          >
-                            <View style={styles.shopCardHeader}>
-                              <View
-                                style={[
-                                  styles.shopIconShell,
-                                  { borderColor: colors.surfaceBorder, backgroundColor: colors.surface },
-                                ]}
-                              >
-                                <LinearGradient
-                                  colors={[colors.sky, colors.emerald]}
-                                  start={{ x: 0, y: 0 }}
-                                  end={{ x: 1, y: 1 }}
-                                  style={styles.shopIconInner}
-                                >
-                                  <MaterialCommunityIcons name={item.icon || 'flash-outline'} size={20} color="#0f172a" />
-                                </LinearGradient>
-                              </View>
-                              <View style={styles.shopCardTitleArea}>
-                                <View style={styles.shopCardTitleRow}>
-                                  <Text style={[styles.shopCardTitle, { color: colors.text }]}>{item.name}</Text>
-                                  {active ? (
-                                    <View
-                                      style={[
-                                        styles.shopCardBadge,
-                                        {
-                                          borderColor: colors.surfaceBorder,
-                                          backgroundColor: hexToRgba(colors.text, eff === 'light' ? 0.05 : 0.22),
-                                        },
-                                      ]}
-                                    >
-                                      <Text style={[styles.shopCardBadgeText, { color: colors.text }]}>In progress</Text>
-                                    </View>
-                                  ) : null}
-                                </View>
-                                <Text style={[styles.shopCardDescription, { color: hexToRgba(colors.text, 0.65) }]}>
-                                  {item.description}
-                                </Text>
-                              </View>
-                            </View>
-                            <View style={styles.shopCardFooter}>
-                              <View style={styles.shopMetaRow}>
-                                <View style={styles.shopMetaItem}>
-                                  <MaterialCommunityIcons name="clock-time-four-outline" size={14} color={hexToRgba(colors.text, 0.6)} />
-                                  <Text style={[styles.shopMetaText, { color: hexToRgba(colors.text, 0.6) }]}>
-                                    {durationLabel}
-                                  </Text>
-                                </View>
-                                <View style={styles.shopMetaItem}>
-                                  <MaterialCommunityIcons name="diamond-stone" size={14} color={hexToRgba(colors.text, 0.6)} />
-                                  <Text style={[styles.shopMetaText, { color: hexToRgba(colors.text, 0.6) }]}>
-                                    {formatGold(item.cost)}
-                                  </Text>
-                                </View>
-                              </View>
-                              <GoldPill
-                                colors={colors}
-                                icon="diamond-stone"
-                                onPress={() => handleBuyEffect(item)}
-                                dim={gold < item.cost || active}
-                              >
-                                {active ? 'Owned' : `${item.cost}g`}
-                              </GoldPill>
-                            </View>
-                          </View>
-                        );
-                      })}
-                  </View>
-                )}
-
-                {categoryTab === 'rewards' && (
-                  <View style={styles.shopCardList}>
-                    {sortedRewards.map((item) => {
-                      const cost = costFor(item);
-                      return (
-                        <View
-                          key={item.id}
-                          style={[
-                            styles.shopCard,
-                            { backgroundColor: filledSurface, borderColor: colors.surfaceBorder },
-                            cardShadow,
-                          ]}
-                        >
-                          <View style={styles.shopCardHeader}>
-                            <View
-                              style={[
-                                styles.shopIconShell,
-                                { borderColor: colors.surfaceBorder, backgroundColor: colors.surface },
-                              ]}
-                            >
-                              <LinearGradient
-                                colors={[colors.rose, colors.lilac]}
-                                start={{ x: 0, y: 0 }}
-                                end={{ x: 1, y: 1 }}
-                                style={styles.shopIconInner}
-                              >
-                                <MaterialCommunityIcons name="gift-outline" size={20} color="#0f172a" />
-                              </LinearGradient>
-                            </View>
-                            <View style={styles.shopCardTitleArea}>
-                              <Text style={[styles.shopCardTitle, { color: colors.text }]}>{item.name}</Text>
-                              <Text style={[styles.shopCardDescription, { color: hexToRgba(colors.text, 0.65) }]}>
-                                Reward yourself with {item.minutes} minutes of joy.
-                              </Text>
-                            </View>
-                          </View>
-                          <View style={styles.shopCardFooter}>
-                            <View style={styles.shopMetaRow}>
-                              <View style={styles.shopMetaItem}>
-                                <MaterialCommunityIcons name="clock-time-four-outline" size={14} color={hexToRgba(colors.text, 0.6)} />
-                                <Text style={[styles.shopMetaText, { color: hexToRgba(colors.text, 0.6) }]}>
-                                  {item.minutes} min
-                                </Text>
-                              </View>
-                              <View style={styles.shopMetaItem}>
-                                <MaterialCommunityIcons name="diamond-stone" size={14} color={hexToRgba(colors.text, 0.6)} />
-                                <Text style={[styles.shopMetaText, { color: hexToRgba(colors.text, 0.6) }]}>
-                                  {formatGold(cost)}
-                                </Text>
-                              </View>
-                            </View>
-                            <GoldPill
-                              colors={colors}
-                              icon="diamond-stone"
-                              onPress={() => setConfirmReward(item)}
-                              dim={gold < cost}
-                            >
-                              {`${cost}g`}
-                            </GoldPill>
-                          </View>
-                        </View>
-                      );
-                    })}
-                  </View>
-                )}
-
-                {categoryTab === 'premium' && (
-                  <View style={styles.shopCardList}>
-                    {PREMIUM_REWARDS.map((item) => {
-                      const cost = costFor(item);
-                      const progress = premiumProgress[item.id] || 0;
-                      const completed = progress >= cost;
-                      const savedGold = Math.min(progress, cost);
-                      const remainingGold = Math.max(cost - savedGold, 0);
-                      const progressPercent = cost > 0 ? Math.min(100, (savedGold / cost) * 100) : 100;
-                      const displayName = (item.name || '').replace(/premium reward/gi, '').trim();
-                      return (
-                        <View
-                          key={item.id}
-                          style={[
-                            styles.shopPremiumCard,
-                            { backgroundColor: filledSurface, borderColor: colors.surfaceBorder },
-                            cardShadow,
-                          ]}
-                        >
-                          <View style={styles.shopPremiumHeader}>
-                            <View
-                              style={[
-                                styles.shopIconShell,
-                                { borderColor: colors.surfaceBorder, backgroundColor: colors.surface },
-                              ]}
-                            >
-                              <LinearGradient
-                                colors={[colors.sky, colors.emerald]}
-                                start={{ x: 0, y: 0 }}
-                                end={{ x: 1, y: 1 }}
-                                style={styles.shopIconInner}
-                              >
-                                <MaterialCommunityIcons name="crown-outline" size={20} color="#0f172a" />
-                              </LinearGradient>
-                            </View>
-                            <View style={styles.shopPremiumTitleArea}>
-                              <View style={styles.shopPremiumTitleRow}>
-                                <Text style={[styles.shopCardTitle, { color: colors.text }]}>
-                                  {displayName || 'Premium reward'}
-                                </Text>
-                                {completed ? (
-                                  <View style={[styles.shopPremiumBadge, { backgroundColor: hexToRgba(colors.emerald, 0.2) }]}>
-                                    <Text style={[styles.shopPremiumBadgeText, { color: colors.emerald }]}>Ready to claim</Text>
-                                  </View>
-                                ) : null}
-                              </View>
-                              <Text style={[styles.shopCardDescription, { color: hexToRgba(colors.text, 0.65) }]}>
-                                Set aside gold to unlock this premium reward.
-                              </Text>
-                            </View>
-                            <GoldPill
-                              colors={colors}
-                              icon="diamond-stone"
-                              onPress={() => handlePremiumAction(item)}
-                              dim={!completed && gold <= 0}
-                              style={styles.shopPremiumButton}
-                            >
-                              {completed ? 'Claim' : 'Save'}
-                            </GoldPill>
-                          </View>
-                          <View
-                            style={[styles.shopPremiumProgressBar, { borderColor: colors.surfaceBorder }]}
-                          >
-                            <LinearGradient
-                              colors={[colors.sky, colors.emerald]}
-                              start={{ x: 0, y: 0 }}
-                              end={{ x: 1, y: 0 }}
-                              style={[styles.shopPremiumProgressFill, { width: `${progressPercent}%` }]}
-                            />
-                          </View>
-                          <View style={styles.shopPremiumAmounts}>
-                            <View style={styles.shopMetaItem}>
-                              <MaterialCommunityIcons name="wallet-outline" size={14} color={hexToRgba(colors.text, 0.6)} />
-                              <Text style={[styles.shopMetaText, { color: hexToRgba(colors.text, 0.6) }]}>
-                                {formatGoldValue(savedGold)}
-                              </Text>
-                            </View>
-                            <Text style={[styles.shopPremiumRemaining, { color: hexToRgba(colors.text, 0.6) }]}>
-                              {remainingGold === 0
-                                ? 'Goal reached'
-                                : `${formatGoldValue(remainingGold)} to go`}
-                            </Text>
-                          </View>
-                        </View>
-                      );
-                    })}
-                  </View>
-                )}
-              </View>
-            )}
-          </View>
-        </View>
-      </ScrollView>
-
-      <Modal
-        visible={!!savingItem}
-        transparent
-        animationType="fade"
-        onRequestClose={closeSavingModal}
-      >
-        <TouchableWithoutFeedback onPress={closeSavingModal}>
-          <View style={[styles.shopModalOverlay, { backgroundColor: overlayBackground }]}>
-            <TouchableWithoutFeedback onPress={() => {}}>
-              <View
-                style={[
-                  styles.shopModalCard,
-                  { backgroundColor: colors.surface, borderColor: colors.surfaceBorder },
-                  cardShadow,
-                ]}
-              >
-                <View style={styles.shopModalHeader}>
-                  <View
-                    style={[
-                      styles.shopModalIconShell,
-                      { borderColor: colors.surfaceBorder, backgroundColor: colors.surface },
-                    ]}
-                  >
-                    <LinearGradient
-                      colors={[colors.sky, colors.emerald]}
-                      start={{ x: 0, y: 0 }}
-                      end={{ x: 1, y: 1 }}
-                      style={styles.shopModalIconInner}
-                    >
-                      <MaterialCommunityIcons name="wallet-outline" size={20} color="#0f172a" />
-                    </LinearGradient>
-                  </View>
-                  <View style={styles.shopModalHeaderText}>
-                    <Text style={[styles.shopModalTitle, { color: colors.text }]}>
-                      Save gold for {savingItem?.name}
-                    </Text>
-                    <Text style={[styles.shopModalDescription, { color: hexToRgba(colors.text, 0.65) }]}>Decide how much to stash away right now. You can come back later to add more.</Text>
-                  </View>
-                </View>
-                <GoldSlider
-                  value={clampedSaveAmount}
-                  max={maxSavableNow}
-                  colors={colors}
-                  eff={eff}
-                  onChange={(next) => setSaveAmount(next)}
-                />
-                <View style={styles.shopModalLabelRow}>
-                  <Text style={[styles.shopModalLabel, { color: colors.text }]}>
-                    {formatGoldValue(clampedSaveAmount)}
-                  </Text>
-                  <Text style={[styles.shopModalLabelValue, { color: hexToRgba(colors.text, 0.6) }]}>
-                    {formatGoldValue(maxSavableNow)} available now
-                  </Text>
-                </View>
-                <Text style={[styles.shopModalHint, { color: hexToRgba(colors.text, 0.6) }]}>
-                  Already saved {formatGoldValue(savingProgress)} / {formatGoldValue(savingItem ? costFor(savingItem) : 0)}
-                </Text>
-                <View style={styles.shopModalActions}>
-                  <TouchableOpacity
-                    onPress={closeSavingModal}
-                    activeOpacity={0.85}
-                    style={[styles.shopModalButton, { borderColor: colors.surfaceBorder, backgroundColor: colors.surface }]}
-                  >
-                    <Text style={[styles.shopModalButtonText, { color: colors.text }]}>Cancel</Text>
-                  </TouchableOpacity>
-                  <TouchableOpacity
-                    onPress={confirmSave}
-                    activeOpacity={clampedSaveAmount > 0 ? 0.85 : 1}
-                    disabled={clampedSaveAmount <= 0}
-                    style={[
-                      styles.shopModalButton,
-                      styles.shopModalPrimary,
-                      {
-                        borderColor: colors.surfaceBorder,
-                        opacity: clampedSaveAmount > 0 ? 1 : 0.5,
-                      },
-                    ]}
-                  >
-                    <LinearGradient
-                      colors={[colors.sky, colors.emerald]}
-                      start={{ x: 0, y: 0 }}
-                      end={{ x: 1, y: 0 }}
-                      style={StyleSheet.absoluteFillObject}
-                    />
-                    <Text style={styles.shopModalPrimaryText}>Save amount</Text>
-                  </TouchableOpacity>
-                </View>
-              </View>
-            </TouchableWithoutFeedback>
-          </View>
-        </TouchableWithoutFeedback>
-      </Modal>
-
-      <Modal
-        visible={!!confirmReward}
-        transparent
-        animationType="fade"
-        onRequestClose={() => setConfirmReward(null)}
-      >
-        <TouchableWithoutFeedback onPress={() => setConfirmReward(null)}>
-          <View style={[styles.shopModalOverlay, { backgroundColor: overlayBackground }]}>
-            <TouchableWithoutFeedback onPress={() => {}}>
-              <View
-                style={[
-                  styles.shopConfirmCard,
-                  { backgroundColor: colors.surface, borderColor: colors.surfaceBorder },
-                  cardShadow,
-                ]}
-              >
-                <View style={styles.shopConfirmIconRow}>
-                  <View
-                    style={[
-                      styles.shopModalIconShell,
-                      { borderColor: colors.surfaceBorder, backgroundColor: colors.surface },
-                    ]}
-                  >
-                    <LinearGradient
-                      colors={[colors.sky, colors.emerald]}
-                      start={{ x: 0, y: 0 }}
-                      end={{ x: 1, y: 1 }}
-                      style={styles.shopModalIconInner}
-                    >
-                      <MaterialCommunityIcons name={ConfirmIcon} size={22} color="#0f172a" />
-                    </LinearGradient>
-                  </View>
-                </View>
-                <Text style={[styles.shopConfirmTitle, { color: colors.text }]}>
-                  {confirmReward?.premium
-                    ? `Redeem ${confirmReward?.name}?`
-                    : `Spend ${formatGold(confirmCost)} for ${confirmReward?.name}?`}
-                </Text>
-                {!confirmReward?.premium && (
-                  <Text style={[styles.shopModalDescription, { color: hexToRgba(colors.text, 0.6), textAlign: 'center' }]}>
-                    {confirmReward?.minutes} minutes of joy await.
-                  </Text>
-                )}
-                <View style={styles.shopModalActions}>
-                  <TouchableOpacity
-                    onPress={() => setConfirmReward(null)}
-                    activeOpacity={0.85}
-                    style={[styles.shopModalButton, { borderColor: colors.surfaceBorder, backgroundColor: colors.surface }]}
-                  >
-                    <Text style={[styles.shopModalButtonText, { color: colors.text }]}>Cancel</Text>
-                  </TouchableOpacity>
-                  <TouchableOpacity
-                    onPress={handleConfirmAction}
-                    activeOpacity={0.9}
-                    style={[styles.shopModalButton, styles.shopModalPrimary, { borderColor: colors.surfaceBorder }]}
-                  >
-                    <LinearGradient
-                      colors={[colors.sky, colors.emerald]}
-                      start={{ x: 0, y: 0 }}
-                      end={{ x: 1, y: 0 }}
-                      style={StyleSheet.absoluteFillObject}
-                    />
-                    <Text style={styles.shopModalPrimaryText}>Yes, do it</Text>
-                  </TouchableOpacity>
-                </View>
-              </View>
-            </TouchableWithoutFeedback>
-          </View>
-        </TouchableWithoutFeedback>
-      </Modal>
-
-      <Modal
-        visible={!!redeemed}
-        transparent
-        animationType="fade"
-        onRequestClose={() => setRedeemed(null)}
-      >
-        <TouchableWithoutFeedback onPress={() => setRedeemed(null)}>
-          <View style={[styles.shopModalOverlay, { backgroundColor: overlayBackground }]}>
-            <TouchableWithoutFeedback onPress={() => {}}>
-              <View
-                style={[
-                  styles.shopConfirmCard,
-                  { backgroundColor: colors.surface, borderColor: colors.surfaceBorder },
-                  cardShadow,
-                ]}
-              >
-                <Text style={[styles.shopConfirmTitle, { color: colors.text }]}>Enjoy {redeemed?.name}!</Text>
-                <Text style={[styles.shopModalDescription, { color: hexToRgba(colors.text, 0.6), textAlign: 'center' }]}>
-                  You earned it. Log it in your journal once you're back.
-                </Text>
-                <TouchableOpacity
-                  onPress={() => setRedeemed(null)}
-                  activeOpacity={0.9}
-                  style={[styles.shopModalButton, styles.shopModalPrimary, { borderColor: colors.surfaceBorder }]}
-                >
-                  <LinearGradient
-                    colors={[colors.sky, colors.emerald]}
-                    start={{ x: 0, y: 0 }}
-                    end={{ x: 1, y: 0 }}
-                    style={StyleSheet.absoluteFillObject}
-                  />
-                  <Text style={styles.shopModalPrimaryText}>Back to shopping</Text>
-                </TouchableOpacity>
-              </View>
-            </TouchableWithoutFeedback>
-          </View>
-        </TouchableWithoutFeedback>
-      </Modal>
-    </>
-  );
-};
-
-
-const ChestCard = ({ chest, colors, theme, isFocused, onFocus, onOpen }) => {
-  const rarity = chest.rarity || 'Common';
-  const ChestArt = CHEST_ART[rarity] || CHEST_ART.Common;
-  const detail = RARITY_DETAILS[rarity] || RARITY_DETAILS.Common;
-  const goldRange = chest.gold ? formatRange(chest.gold) : '?';
-  const cardBackground = theme === 'light' ? ensureOpaque(colors.surface) : 'rgba(15,23,42,0.55)';
-  const iconBackground = theme === 'light' ? ensureOpaque(colors.surface) : 'rgba(15,23,42,0.45)';
-
-  const overlayColors =
-    theme === 'light'
-      ? ['rgba(255,255,255,0.96)', 'rgba(226,232,240,0.86)']
-      : ['rgba(15,23,42,0.9)', 'rgba(15,23,42,0.78)'];
-
-  return (
-    <TouchableOpacity
-      activeOpacity={0.9}
-      onPress={() => onFocus(isFocused ? null : chest.id)}
-      style={[
-        styles.chestCard,
-        {
-          backgroundColor: cardBackground,
-          borderColor: colors.surfaceBorder,
-        },
-        theme === 'light' ? styles.chestCardShadowLight : styles.chestCardShadowDark,
-      ]}
-    >
-      <View style={styles.chestCardInner}>
-        <View
-          style={[
-            styles.chestIconWrapper,
-            {
-              backgroundColor: iconBackground,
-              borderColor: colors.surfaceBorder,
-            },
-          ]}
-        >
-          <ChestArt size={54} />
-        </View>
-        <Text style={[styles.chestRarity, { color: hexToRgba(colors.text, 0.7) }]}>{rarity} chest</Text>
-        <Text style={[styles.chestHeadline, { color: hexToRgba(colors.text, 0.82) }]}>{detail.headline}</Text>
-      </View>
-      {isFocused ? (
-        <LinearGradient colors={overlayColors} start={{ x: 0, y: 0 }} end={{ x: 1, y: 1 }} style={styles.chestOverlay}>
-          <Text
-            style={[
-              styles.chestOverlayTitle,
-              { color: theme === 'light' ? 'rgba(15,23,42,0.68)' : 'rgba(226,232,240,0.78)' },
-            ]}
-          >
-            Gold stash
-          </Text>
-          <Text style={[styles.chestOverlayRange, { color: theme === 'light' ? '#0f172a' : colors.text }]}>
-            {goldRange}g
-          </Text>
-          <Text
-            style={[
-              styles.chestOverlayHelper,
-              { color: theme === 'light' ? 'rgba(15,23,42,0.65)' : 'rgba(226,232,240,0.78)' },
-            ]}
-          >
-            {detail.helper}
-          </Text>
-          <TouchableOpacity
-            onPress={() => {
-              const reward = onOpen?.(chest);
-              if (reward) {
-                onFocus(null);
-              }
-            }}
-            activeOpacity={0.9}
-            style={[styles.chestOpenButton, { borderColor: colors.surfaceBorder }]}
-          >
-            <LinearGradient
-              colors={[colors.sky, colors.emerald]}
-              start={{ x: 0, y: 0 }}
-              end={{ x: 1, y: 0 }}
-              style={styles.chestOpenButtonBg}
-            >
-              <MaterialCommunityIcons name="auto-fix" size={16} color="#0f172a" />
-              <Text style={styles.chestOpenButtonText}>Open chest</Text>
-            </LinearGradient>
-          </TouchableOpacity>
-        </LinearGradient>
-      ) : null}
-    </TouchableOpacity>
-  );
-};
-
-const MONTHS = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
-
-const MIN_FOCUS_FOR_LOG = 0.25;
 export default function App() {
   const [fontsLoaded, fontError] = useFonts(MaterialCommunityIcons.font);
   const { mode, eff, cycle } = useTheme();
@@ -1564,16 +185,51 @@ export default function App() {
   const [manualLogs, setManualLogs] = useState(() => ({}));
   const [eventStates, setEventStates] = useState(() => ({}));
   const [currentTime, setCurrentTime] = useState(() => Date.now());
-  const [chests, setChests] = useState([]);
-  const [chestFilter, setChestFilter] = useState('All');
-  const [focusedChestId, setFocusedChestId] = useState(null);
-  const [openAllSummary, setOpenAllSummary] = useState(null);
-  const [openResult, setOpenResult] = useState(null);
   const [premiumProgress, setPremiumProgress] = useState({});
   const [isHydrated, setIsHydrated] = useState(false);
   const [hydrationError, setHydrationError] = useState(null);
   const [shopMainTab, setShopMainTab] = useState('catalogue');
   const [shopCategoryTab, setShopCategoryTab] = useState('effects');
+
+  const goldMultiplier = useMemo(
+    () => (activeEffects.some((effect) => effect.id === 2) ? 2 : 1),
+    [activeEffects],
+  );
+
+  const {
+    chests,
+    setChests,
+    chestFilter,
+    setChestFilter,
+    focusedChestId,
+    setFocusedChestId,
+    openChest,
+    openAll,
+    openAllSummary,
+    dismissOpenAllSummary,
+    openResult,
+    dismissOpenResult,
+    rarityKeys,
+    rarityCounts,
+    visibleChests,
+    hasChests,
+    viewRange,
+  } = useRewardsInventory({ goldMultiplier, setGold });
+
+  const handleSelectChestFilter = useCallback(
+    (key) => {
+      setChestFilter(key);
+      setFocusedChestId(null);
+    },
+    [setChestFilter, setFocusedChestId],
+  );
+
+  useEffect(() => {
+    if (activeTab !== 'Rewards') {
+      setFocusedChestId(null);
+      dismissOpenResult();
+    }
+  }, [activeTab, dismissOpenResult, setFocusedChestId]);
 
   const {
     appsQuery,
@@ -1589,7 +245,6 @@ export default function App() {
   } = useApplicationFilters(applications);
 
   const lowQualityStreakRef = useRef(0);
-  const openResultTimer = useRef(null);
   const effectWarningTimers = useRef(new Map());
   const announcedEffectKeysRef = useRef(new Set());
   const eventSeenRef = useRef(new Map());
@@ -1874,25 +529,16 @@ export default function App() {
     return () => clearInterval(interval);
   }, [hasTimedEffects]);
 
-  const sprayActive = useMemo(() => {
-    if (!sprayDebuff) {
-      return false;
-    }
-    const { expiresAt } = sprayDebuff;
-    if (!expiresAt) {
-      return true;
-    }
-    return expiresAt > currentTime;
-  }, [sprayDebuff, currentTime]);
+  const sprayActive = useMemo(
+    () => isSprayDebuffActive(sprayDebuff, currentTime),
+    [sprayDebuff, currentTime],
+  );
 
   useEffect(() => {
-    if (!sprayDebuff) {
-      return;
-    }
-    if (sprayDebuff.expiresAt && sprayDebuff.expiresAt <= currentTime) {
+    if (shouldClearSprayDebuff(sprayDebuff, currentTime)) {
       setSprayDebuff(null);
     }
-  }, [sprayDebuff, currentTime]);
+  }, [sprayDebuff, currentTime, setSprayDebuff]);
 
   const sprayMultiplier = sprayActive ? 0.5 : 1;
 
@@ -1900,26 +546,13 @@ export default function App() {
     if (!sprayActive) {
       return null;
     }
-    const expiresAt =
-      sprayDebuff?.expiresAt ??
-      (sprayDebuff?.activatedAt ? sprayDebuff.activatedAt + SPRAY_DEBUFF_DURATION_MS : undefined);
-    return {
-      id: 'spray-and-pray',
-      name: 'Spray and Pray',
-      icon: 'spray-bottle',
-      description: 'XP and gold from applications are halved for six hours.',
-      expiresAt,
-      duration: SPRAY_DEBUFF_DURATION_SECONDS,
-      type: 'debuff',
-    };
+    return createSprayEffectDetails(sprayDebuff);
   }, [sprayActive, sprayDebuff]);
 
-  const shopEffects = useMemo(() => {
-    if (!sprayEffectDetails) {
-      return activeEffects;
-    }
-    return [sprayEffectDetails, ...activeEffects];
-  }, [sprayEffectDetails, activeEffects]);
+  const shopEffects = useMemo(
+    () => includeSprayEffect(activeEffects, sprayEffectDetails),
+    [activeEffects, sprayEffectDetails],
+  );
 
   useEffect(() => {
     const previous = announcedEffectKeysRef.current;
@@ -2099,42 +732,6 @@ export default function App() {
     }
     setShowForm(true);
   }, [focus]);
-
-  const openChest = useCallback(
-    (chest) => {
-      if (!chest) {
-        return null;
-      }
-      const base = rand(chest.gold);
-      const totalGold = Math.round(base * goldMultiplier);
-      setGold((value) => value + totalGold);
-      setChests((prev) => prev.filter((item) => item.id !== chest.id));
-      setFocusedChestId(null);
-      setOpenResult({ gold: totalGold });
-      if (openResultTimer.current) {
-        clearTimeout(openResultTimer.current);
-      }
-      openResultTimer.current = setTimeout(() => {
-        setOpenResult(null);
-      }, 1600);
-      return { gold: totalGold };
-    },
-    [goldMultiplier],
-  );
-
-  const openAll = useCallback(() => {
-    if (!chests.length) {
-      return;
-    }
-    let totalGold = 0;
-    chests.forEach((item) => {
-      totalGold += Math.round(rand(item.gold) * goldMultiplier);
-    });
-    setGold((value) => value + totalGold);
-    setChests([]);
-    setFocusedChestId(null);
-    setOpenAllSummary({ gold: totalGold, opened: chests.length });
-  }, [chests, goldMultiplier]);
 
   const { statsSnapshot, weeklyTrend } = useMemo(() => {
     const now = Number.isFinite(currentTime) ? currentTime : Date.now();
@@ -2332,13 +929,6 @@ export default function App() {
       default: fallback,
     };
   }, [colors, eff]);
-
-  const totalPotential = useMemo(() => computePotential(chests), [chests]);
-  const viewRange = totalPotential ? `${formatRange(totalPotential)}g` : '0g';
-  const hasChests = chests.length > 0;
-  const summaryMuted = eff === 'light' ? 'rgba(15,23,42,0.66)' : 'rgba(226,232,240,0.76)';
-  const summaryStrong = eff === 'light' ? '#0f172a' : colors.text;
-  const headlineColor = hexToRgba(colors.text, 0.85);
 
   const statusIcons = useMemo(() => {
     const entries = {};
@@ -2564,11 +1154,6 @@ export default function App() {
     [colors],
   );
 
-  const goldMultiplier = useMemo(
-    () => (activeEffects.some((effect) => effect.id === 2) ? 2 : 1),
-    [activeEffects],
-  );
-
   const applyRewardEffect = useCallback(
     (label) => {
       if (!label || typeof label !== 'string') {
@@ -2691,14 +1276,6 @@ export default function App() {
     },
     [gainXp, setGold, applyRewardEffect, setEventStates, updateCurrentTime],
   );
-
-  useEffect(() => {
-    return () => {
-      if (openResultTimer.current) {
-        clearTimeout(openResultTimer.current);
-      }
-    };
-  }, []);
 
   useEffect(() => {
     return () => {
@@ -3102,205 +1679,44 @@ export default function App() {
             contentContainerStyle={styles.scrollContent}
             showsVerticalScrollIndicator={false}
           >
-            <LinearGradient
-              colors={[hexToRgba(colors.sky, 0.28), hexToRgba(colors.emerald, 0.22)]}
-              start={{ x: 0, y: 0 }}
-              end={{ x: 1, y: 1 }}
-              style={[styles.rewardsSummary, { borderColor: colors.surfaceBorder }]}
-            >
-              <View style={styles.rewardsSummaryHeader}>
-                <View style={styles.rewardsSummaryTitle}>
-                  <MaterialCommunityIcons name="gift-outline" size={16} color={headlineColor} />
-                  <Text style={[styles.rewardsSummaryTitleText, { color: headlineColor }]}>Treasure vault</Text>
-                </View>
-                <TouchableOpacity
-                  onPress={openAll}
-                  disabled={!hasChests}
-                  activeOpacity={hasChests ? 0.9 : 1}
-                  style={[
-                    styles.rewardsSummaryButton,
-                    {
-                      borderColor: hasChests ? hexToRgba(colors.emerald, 0.45) : colors.surfaceBorder,
-                      backgroundColor: hasChests
-                        ? 'transparent'
-                        : hexToRgba(colors.text, eff === 'light' ? 0.08 : 0.18),
-                    },
-                  ]}
-                >
-                  {hasChests ? (
-                    <LinearGradient
-                      colors={[colors.sky, colors.emerald]}
-                      start={{ x: 0, y: 0 }}
-                      end={{ x: 1, y: 0 }}
-                      style={styles.rewardsSummaryButtonBackground}
-                    />
-                  ) : null}
-                  <View style={styles.rewardsSummaryButtonContent}>
-                    <MaterialCommunityIcons
-                      name="auto-fix"
-                      size={14}
-                      color={hasChests ? '#0f172a' : hexToRgba(colors.text, 0.55)}
-                    />
-                    <Text
-                      style={[
-                        styles.rewardsSummaryButtonText,
-                        { color: hasChests ? '#0f172a' : hexToRgba(colors.text, 0.55) },
-                      ]}
-                    >
-                      Open all
-                    </Text>
-                  </View>
-                </TouchableOpacity>
-              </View>
-
-              <View style={styles.rewardsSummaryRow}>
-                <View style={styles.rewardsSummaryStat}>
-                  <MaterialCommunityIcons name="diamond-stone" size={14} color={summaryMuted} />
-                  <Text style={[styles.rewardsSummaryStatText, { color: summaryMuted }]}>Potential</Text>
-                  <Text style={[styles.rewardsSummaryRange, { color: summaryStrong }]}>{viewRange}</Text>
-                </View>
-              </View>
-            </LinearGradient>
-
-            <Text style={[styles.rewardsFilterLabel, { color: hexToRgba(colors.text, 0.6) }]}>Choose rarity</Text>
-            <ScrollView
-              horizontal
-              showsHorizontalScrollIndicator={false}
-              contentContainerStyle={styles.rewardsFilterRow}
-            >
-              {rarityKeys.map((key) => {
-                const active = chestFilter === key;
-                return (
-                  <TouchableOpacity
-                    key={key}
-                    onPress={() => {
-                      setChestFilter(key);
-                      setFocusedChestId(null);
-                    }}
-                    activeOpacity={0.9}
-                    style={[
-                      styles.rarityButton,
-                      {
-                        borderColor: active
-                          ? hexToRgba(colors.emerald, 0.45)
-                          : hexToRgba('#0f172a', eff === 'light' ? 0.08 : 0.24),
-                        backgroundColor: active
-                          ? 'transparent'
-                          : hexToRgba('#0f172a', eff === 'light' ? 0.04 : 0.32),
-                      },
-                    ]}
-                  >
-                    {active ? (
-                      <LinearGradient
-                        colors={[colors.sky, colors.emerald]}
-                        start={{ x: 0, y: 0 }}
-                        end={{ x: 1, y: 0 }}
-                        style={styles.rarityButtonBackground}
-                      />
-                    ) : null}
-                    <View style={styles.rarityButtonContent}>
-                      <Text style={[styles.rarityButtonText, { color: active ? '#0f172a' : colors.text }]}>
-                        {key}
-                      </Text>
-                      <View
-                        style={[
-                          styles.rarityBadge,
-                          {
-                            backgroundColor: active
-                              ? 'rgba(255,255,255,0.6)'
-                              : hexToRgba('#0f172a', eff === 'light' ? 0.08 : 0.28),
-                          },
-                        ]}
-                      >
-                        <Text
-                          style={[
-                            styles.rarityBadgeText,
-                            { color: active ? '#0f172a' : hexToRgba(colors.text, 0.7) },
-                          ]}
-                        >
-                          {rarityCounts[key] || 0}
-                        </Text>
-                      </View>
-                    </View>
-                  </TouchableOpacity>
-                );
-              })}
-            </ScrollView>
-
-            <View style={styles.chestGrid}>
-              {visibleChests.map((chest) => (
-                <ChestCard
-                  key={chest.id}
-                  chest={chest}
+            <View style={styles.rewardsSection}>
+              <View style={styles.rewardsSummaryWrapper}>
+                <RewardSummary
                   colors={colors}
-                  theme={eff}
-                  isFocused={focusedChestId === chest.id}
-                  onFocus={setFocusedChestId}
-                  onOpen={openChest}
+                  eff={eff}
+                  hasChests={hasChests}
+                  viewRange={viewRange}
+                  onOpenAll={openAll}
+                  openAllSummary={openAllSummary}
+                  onCloseSummary={dismissOpenAllSummary}
                 />
-              ))}
-              {!visibleChests.length && (
-                <LinearGradient
-                  colors={[hexToRgba(colors.sky, 0.18), hexToRgba(colors.emerald, 0.12)]}
-                  start={{ x: 0, y: 0 }}
-                  end={{ x: 1, y: 1 }}
-                  style={[styles.chestEmpty, { borderColor: colors.surfaceBorder }]}
-                >
-                  <Text style={[styles.chestEmptyText, { color: hexToRgba(colors.text, 0.7) }]}>
-                    No chests match this view.
-                  </Text>
-                </LinearGradient>
-              )}
+              </View>
+              <ChestGrid
+                rarityKeys={rarityKeys}
+                rarityCounts={rarityCounts}
+                chestFilter={chestFilter}
+                onSelectFilter={handleSelectChestFilter}
+                colors={colors}
+                eff={eff}
+                visibleChests={visibleChests}
+                focusedChestId={focusedChestId}
+                onFocus={setFocusedChestId}
+                onOpen={openChest}
+              />
             </View>
           </ScrollView>
 
-          <Modal
-            visible={!!openAllSummary}
-            transparent
-            animationType="fade"
-            onRequestClose={() => setOpenAllSummary(null)}
-          >
-            <View style={styles.rewardsModalOverlay}>
-              <View
-                style={[
-                  styles.rewardsModalCard,
-                  { backgroundColor: colors.surface, borderColor: colors.surfaceBorder },
-                  eff === 'light' ? styles.chestCardShadowLight : styles.chestCardShadowDark,
-                ]}
-              >
-                <Text style={[styles.rewardsModalTitle, { color: colors.text }]}> 
-                  Opened {openAllSummary?.opened} chest{openAllSummary?.opened === 1 ? '' : 's'}
-                </Text>
-                <View style={styles.rewardsModalGold}>
-                  <MaterialCommunityIcons name="diamond-stone" size={18} color={colors.emerald} />
-                  <Text style={[styles.rewardsModalGoldText, { color: colors.text }]}>
-                    {openAllSummary?.gold}g
-                  </Text>
-                </View>
-                <TouchableOpacity
-                  onPress={() => setOpenAllSummary(null)}
-                  activeOpacity={0.9}
-                  style={[styles.rewardsModalButton, { borderColor: colors.surfaceBorder }]}
-                >
-                  <LinearGradient
-                    colors={[colors.sky, colors.emerald]}
-                    start={{ x: 0, y: 0 }}
-                    end={{ x: 1, y: 0 }}
-                    style={styles.rewardsModalButtonBackground}
-                  >
-                    <Text style={styles.rewardsModalButtonText}>Nice!</Text>
-                  </LinearGradient>
-                </TouchableOpacity>
-              </View>
-            </View>
-          </Modal>
-
           {openResult ? (
-            <ToastBanner colors={colors} eff={eff} title="Fresh loot" icon="diamond-stone" value={`${openResult.gold}g`} />
+            <ToastBanner
+              colors={colors}
+              eff={eff}
+              title="Fresh loot"
+              icon="diamond-stone"
+              value={`${openResult.gold}g`}
+            />
           ) : null}
         </>
       )}
-
       {activeTab === 'Shop' && (
         <ShopScreen
           colors={colors}
@@ -4295,755 +2711,10 @@ submitButton: {
     borderRadius: 12,
     alignItems: 'center',
   },
-rewardsSummary: {
-    borderRadius: 24,
-    padding: 20,
-    marginBottom: 20,
-    borderWidth: 1,
-  },
-  rewardsSummaryHeader: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginBottom: 16,
-  },
-  rewardsSummaryTitle: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 8,
-  },
-  rewardsSummaryTitleText: {
-    fontSize: 13,
-    fontWeight: '700',
-    textTransform: 'uppercase',
-    letterSpacing: 0.6,
-  },
-  rewardsSummaryButton: {
-    borderWidth: 1,
-    borderRadius: 999,
-    paddingHorizontal: 16,
-    paddingVertical: 10,
-    overflow: 'hidden',
-  },
-  rewardsSummaryButtonBackground: {
-    position: 'absolute',
-    top: 0,
-    right: 0,
-    bottom: 0,
-    left: 0,
-    borderRadius: 999,
-  },
-  rewardsSummaryButtonContent: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center',
-    gap: 8,
-  },
-  rewardsSummaryButtonText: {
-    fontSize: 12,
-    fontWeight: '600',
-  },
-  rewardsSummaryRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    flexWrap: 'wrap',
-    gap: 12,
-  },
-  rewardsSummaryStat: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 8,
-  },
-  rewardsSummaryStatText: {
-    fontSize: 12,
-    fontWeight: '600',
-  },
-  rewardsSummaryRange: {
-    fontSize: 13,
-    fontWeight: '700',
-  },
-  rewardsFilterLabel: {
-    fontSize: 11,
-    fontWeight: '600',
-    textTransform: 'uppercase',
-    letterSpacing: 0.7,
-    marginBottom: 10,
-  },
-  rewardsFilterRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 12,
-    paddingBottom: 6,
-  },
-  rarityButton: {
-    borderWidth: 1,
-    borderRadius: 999,
-    paddingHorizontal: 16,
-    paddingVertical: 10,
-    overflow: 'hidden',
-  },
-  rarityButtonBackground: {
-    position: 'absolute',
-    top: 0,
-    right: 0,
-    bottom: 0,
-    left: 0,
-    borderRadius: 999,
-  },
-  rarityButtonContent: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 8,
-  },
-  rarityButtonText: {
-    fontSize: 12,
-    fontWeight: '600',
-  },
-  rarityBadge: {
-    borderRadius: 999,
-    paddingHorizontal: 8,
-    paddingVertical: 2,
-  },
-  rarityBadgeText: {
-    fontSize: 11,
-    fontWeight: '600',
-  },
-  chestGrid: {
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-    justifyContent: 'space-between',
-    gap: 12,
-    marginTop: 16,
-  },
-  chestCard: {
-    width: '48%',
-    borderRadius: 20,
-    borderWidth: 1,
-    paddingVertical: 16,
-    paddingHorizontal: 12,
-    marginBottom: 12,
-    overflow: 'hidden',
-  },
-  chestCardInner: {
-    alignItems: 'center',
-    gap: 12,
-  },
-  chestIconWrapper: {
-    width: 68,
-    height: 68,
-    borderRadius: 18,
-    alignItems: 'center',
-    justifyContent: 'center',
-    borderWidth: 1,
-  },
-  chestRarity: {
-    fontSize: 12,
-    fontWeight: '700',
-    textTransform: 'uppercase',
-  },
-  chestHeadline: {
-    fontSize: 14,
-    fontWeight: '600',
-    textAlign: 'center',
-  },
-  chestOverlay: {
-    position: 'absolute',
-    top: 0,
-    right: 0,
-    bottom: 0,
-    left: 0,
-    alignItems: 'center',
-    justifyContent: 'center',
-    paddingHorizontal: 16,
-    paddingVertical: 18,
-    gap: 8,
-  },
-  chestOverlayTitle: {
-    fontSize: 11,
-    fontWeight: '600',
-    textTransform: 'uppercase',
-    letterSpacing: 0.6,
-  },
-  chestOverlayRange: {
-    fontSize: 18,
-    fontWeight: '700',
-  },
-  chestOverlayHelper: {
-    fontSize: 12,
-    textAlign: 'center',
-    lineHeight: 16,
-  },
-  chestOpenButton: {
-    borderWidth: 1,
-    borderRadius: 999,
-    overflow: 'hidden',
-    marginTop: 8,
-  },
-  chestOpenButtonBg: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center',
-    gap: 8,
-    paddingHorizontal: 18,
-    paddingVertical: 10,
-  },
-  chestOpenButtonText: {
-    fontSize: 14,
-    fontWeight: '600',
-    color: '#0f172a',
-  },
-  chestEmpty: {
-    width: '100%',
-    borderRadius: 24,
-    borderWidth: 1,
-    paddingVertical: 32,
-    paddingHorizontal: 24,
-    alignItems: 'center',
-    justifyContent: 'center',
-    marginTop: 8,
-  },
-  chestEmptyText: {
-    fontSize: 14,
-    fontWeight: '600',
-    textAlign: 'center',
-  },
-  shopContainer: {
-    borderRadius: 24,
-    borderWidth: 1,
-    padding: 20,
-    marginBottom: 24,
-    overflow: 'hidden',
-  },
-  shopGlowTopLeft: {
-    position: 'absolute',
-    top: -120,
-    left: -120,
-    width: 240,
-    height: 240,
-    borderRadius: 240,
-  },
-  shopGlowTopRight: {
-    position: 'absolute',
-    top: -120,
-    right: -120,
-    width: 240,
-    height: 240,
-    borderRadius: 240,
-  },
-  shopGlowBottomRight: {
-    position: 'absolute',
-    bottom: -140,
-    right: -120,
-    width: 260,
-    height: 260,
-    borderRadius: 260,
-  },
-  shopGlowBottomLeft: {
-    position: 'absolute',
-    bottom: -140,
-    left: -120,
-    width: 260,
-    height: 260,
-    borderRadius: 260,
-  },
-  shopContent: {
-    gap: 20,
-  },
-  shopMainTabs: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    borderWidth: 1,
-    borderRadius: 999,
-    padding: 4,
-    gap: 6,
-  },
-  shopMainTabButton: {
-    flex: 1,
-    borderRadius: 999,
-    borderWidth: 1,
-    overflow: 'hidden',
-  },
-  shopMainTabContent: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center',
-    gap: 6,
-    paddingVertical: 10,
-    paddingHorizontal: 12,
-  },
-  shopMainTabIcon: {
-    marginRight: 0,
-  },
-  shopMainTabLabel: {
-    fontSize: 12,
-    fontWeight: '600',
-  },
-  shopSection: {
-    marginTop: 16,
-    gap: 18,
-  },
-  shopSectionHeader: {
-    gap: 4,
-  },
-  shopSectionEyebrow: {
-    fontSize: 11,
-    fontWeight: '700',
-    textTransform: 'uppercase',
-    letterSpacing: 0.8,
-  },
-  shopSectionTitle: {
-    fontSize: 14,
-    fontWeight: '600',
-  },
-  shopEmptyCard: {
-    borderRadius: 20,
-    borderWidth: 1,
-    padding: 20,
-    gap: 12,
-  },
-  shopEmptyHeader: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 8,
-  },
-  shopEmptyTitle: {
-    fontSize: 14,
-    fontWeight: '600',
-  },
-  shopEmptyDescription: {
-    fontSize: 12,
-    lineHeight: 18,
-  },
-  shopBrowseButton: {
-    marginTop: 8,
-    alignSelf: 'flex-start',
-    borderRadius: 12,
-    borderWidth: 1,
-    overflow: 'hidden',
-  },
-  shopBrowseButtonText: {
-    fontSize: 13,
-    fontWeight: '600',
-    color: '#0f172a',
-    paddingHorizontal: 16,
-    paddingVertical: 10,
-  },
-  shopActiveList: {
-    flexDirection: 'column',
-    gap: 12,
-  },
-  shopActiveCard: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    borderWidth: 1,
-    borderRadius: 20,
-    padding: 14,
-    gap: 12,
-    width: '100%',
-  },
-  shopActiveIconShell: {
-    width: 56,
-    height: 56,
-    borderRadius: 28,
-    borderWidth: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  shopActiveIconInner: {
-    width: '100%',
-    height: '100%',
-    borderRadius: 28,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  shopActiveInfo: {
-    flex: 1,
-    gap: 6,
-  },
-  shopActiveNameRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  shopActiveName: {
-    fontSize: 13,
-    fontWeight: '600',
-  },
-  shopActiveTimerRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 4,
-  },
-  shopActiveTimerText: {
-    fontSize: 11,
-    fontWeight: '600',
-  },
-  shopActivePassive: {
-    fontSize: 11,
-    fontWeight: '600',
-    textTransform: 'uppercase',
-  },
-  shopActiveDescription: {
-    fontSize: 11,
-    lineHeight: 16,
-  },
-  shopSecondaryTabs: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    borderWidth: 1,
-    borderRadius: 999,
-    padding: 4,
-    gap: 6,
-  },
-  shopSecondaryTabButton: {
-    flex: 1,
-    borderRadius: 999,
-    borderWidth: 1,
-    overflow: 'hidden',
-  },
-  shopSecondaryTabContent: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center',
-    gap: 6,
-    paddingVertical: 9,
-    paddingHorizontal: 12,
-  },
-  shopSecondaryTabIcon: {
-    marginRight: 0,
-  },
-  shopSecondaryTabLabel: {
-    fontSize: 12,
-    fontWeight: '600',
-  },
-  shopCardList: {
-    gap: 12,
-  },
-  shopCard: {
-    borderRadius: 20,
-    borderWidth: 1,
-    padding: 16,
-    gap: 16,
-  },
-  shopCardHeader: {
-    flexDirection: 'row',
-    alignItems: 'flex-start',
-    gap: 12,
-  },
-  shopIconShell: {
-    width: 48,
-    height: 48,
-    borderRadius: 16,
-    borderWidth: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  shopIconInner: {
-    width: '100%',
-    height: '100%',
-    borderRadius: 16,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  shopCardTitleArea: {
-    flex: 1,
-    gap: 6,
-  },
-  shopCardTitleRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 8,
-    flexWrap: 'wrap',
-  },
-  shopCardTitle: {
-    fontSize: 14,
-    fontWeight: '700',
-  },
-  shopCardBadge: {
-    borderWidth: 1,
-    borderRadius: 12,
-    paddingHorizontal: 8,
-    paddingVertical: 4,
-  },
-  shopCardBadgeText: {
-    fontSize: 10,
-    fontWeight: '600',
-  },
-  shopCardDescription: {
-    fontSize: 12,
-    lineHeight: 18,
-  },
-  shopCardFooter: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-  },
-  shopMetaRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 12,
-  },
-  shopMetaItem: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 4,
-  },
-  shopMetaText: {
-    fontSize: 12,
-    fontWeight: '500',
-  },
-  shopPremiumCard: {
-    borderRadius: 22,
-    borderWidth: 1,
-    padding: 18,
-    gap: 16,
-  },
-  shopPremiumHeader: {
-    flexDirection: 'row',
-    alignItems: 'flex-start',
-    gap: 12,
-  },
-  shopPremiumTitleArea: {
-    flex: 1,
-    gap: 6,
-  },
-  shopPremiumTitleRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 8,
-    flexWrap: 'wrap',
-  },
-  shopPremiumBadge: {
-    paddingHorizontal: 10,
-    paddingVertical: 4,
-    borderRadius: 999,
-  },
-  shopPremiumBadgeText: {
-    fontSize: 10,
-    fontWeight: '700',
-  },
-  shopPremiumButton: {
-    marginLeft: 'auto',
-  },
-  shopPremiumProgressBar: {
-    height: 10,
-    borderRadius: 999,
-    borderWidth: 1,
-    overflow: 'hidden',
-    marginTop: 12,
-  },
-  shopPremiumProgressFill: {
-    height: '100%',
-    borderRadius: 999,
-  },
-  shopPremiumAmounts: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    marginTop: 12,
-  },
-  shopPremiumRemaining: {
-    fontSize: 11,
-    fontWeight: '600',
-  },
-  shopSliderTrack: {
-    marginTop: 16,
-    marginBottom: 12,
-    height: 14,
-    borderRadius: 999,
-    borderWidth: 1,
-    justifyContent: 'center',
-  },
-  shopSliderFill: {
-    height: '100%',
-    borderRadius: 999,
-  },
-  shopSliderThumb: {
-    position: 'absolute',
-    top: -4,
-    borderWidth: 1,
-    borderRadius: 12,
-  },
-  shopModalOverlay: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    paddingHorizontal: 24,
-  },
-  shopModalCard: {
-    width: '100%',
-    maxWidth: 360,
-    borderRadius: 24,
-    borderWidth: 1,
-    padding: 24,
-    gap: 20,
-  },
-  shopModalHeader: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 12,
-  },
-  shopModalIconShell: {
-    width: 44,
-    height: 44,
-    borderRadius: 14,
-    borderWidth: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  shopModalIconInner: {
-    width: '100%',
-    height: '100%',
-    borderRadius: 14,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  shopModalHeaderText: {
-    flex: 1,
-    gap: 6,
-  },
-  shopModalTitle: {
-    fontSize: 15,
-    fontWeight: '700',
-  },
-  shopModalDescription: {
-    fontSize: 12,
-    lineHeight: 18,
-  },
-  shopModalLabelRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    marginTop: 12,
-  },
-  shopModalLabel: {
-    fontSize: 14,
-    fontWeight: '700',
-  },
-  shopModalLabelValue: {
-    fontSize: 12,
-  },
-  shopModalHint: {
-    fontSize: 11,
-  },
-  shopModalActions: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 12,
-    marginTop: 16,
-  },
-  shopModalButton: {
-    flex: 1,
-    borderRadius: 12,
-    borderWidth: 1,
-    paddingVertical: 12,
-    alignItems: 'center',
-    justifyContent: 'center',
-    overflow: 'hidden',
-  },
-  shopModalButtonText: {
-    fontSize: 13,
-    fontWeight: '600',
-  },
-  shopModalPrimary: {
-    backgroundColor: 'transparent',
-  },
-  shopModalPrimaryText: {
-    fontSize: 13,
-    fontWeight: '700',
-    color: '#0f172a',
-  },
-  shopConfirmCard: {
-    width: '100%',
-    maxWidth: 320,
-    borderRadius: 24,
-    borderWidth: 1,
-    padding: 24,
-    alignItems: 'center',
-    gap: 16,
-  },
-  shopConfirmIconRow: {
-    alignItems: 'center',
-  },
-  shopConfirmTitle: {
-    fontSize: 15,
-    fontWeight: '700',
-    textAlign: 'center',
-  },
-  shopCardShadowLight: {
-    shadowColor: '#0f172a',
-    shadowOpacity: 0.14,
-    shadowRadius: 18,
-    shadowOffset: { width: 0, height: 10 },
-    elevation: 5,
-  },
-  shopCardShadowDark: {
-    shadowColor: '#000',
-    shadowOpacity: 0.45,
-    shadowRadius: 20,
-    shadowOffset: { width: 0, height: 12 },
-    elevation: 7,
-  },
-
-  rewardsModalOverlay: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'rgba(0,0,0,0.45)',
-    paddingHorizontal: 24,
-  },
-  rewardsModalCard: {
-    width: '100%',
-    maxWidth: 320,
-    borderRadius: 24,
-    borderWidth: 1,
-    padding: 24,
-    alignItems: 'center',
-  },
-  rewardsModalTitle: {
-    fontSize: 15,
-    fontWeight: '700',
-    marginBottom: 12,
-    textAlign: 'center',
-  },
-  rewardsModalGold: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 8,
-    marginBottom: 16,
-  },
-  rewardsModalGoldText: {
-    fontSize: 18,
-    fontWeight: '700',
-  },
-  rewardsModalButton: {
-    borderWidth: 1,
-    borderRadius: 12,
-    overflow: 'hidden',
-  },
-  rewardsModalButtonBackground: {
-    paddingHorizontal: 24,
-    paddingVertical: 12,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  rewardsModalButtonText: {
-    fontSize: 14,
-    fontWeight: '700',
-    color: '#0f172a',
-  },
-  chestCardShadowLight: {
-    shadowColor: '#0f172a',
-    shadowOpacity: 0.1,
-    shadowRadius: 14,
-    shadowOffset: { width: 0, height: 10 },
-    elevation: 4,
-  },
-  chestCardShadowDark: {
-    shadowColor: '#000',
-    shadowOpacity: 0.4,
-    shadowRadius: 18,
-    shadowOffset: { width: 0, height: 10 },
-    elevation: 8,
+  rewardsSection: {
+    paddingBottom: 24,
+  },
+  rewardsSummaryWrapper: {
+    marginBottom: 18,
   },
 });
