+++ src/features/rewards/index.js
@@ -1,1 +1,6 @@
-export default {};
+export { default as useRewardsInventory } from './hooks/useRewardsInventory';
+export * from './hooks/useRewardsInventory';
+export { default as RewardSummary } from './components/RewardSummary';
+export { default as ChestGrid } from './components/ChestGrid';
+export { default as ChestCard } from './components/ChestCard';
+export * from './utils/chests';
