+++ src/features/shop/index.js
@@ -1,1 +1,5 @@
-export default {};
+export { default as ShopScreen } from './components/ShopScreen';
+export { default as ShopActiveEffects } from './components/ShopActiveEffects';
+export { default as ShopCatalogue, SHOP_CATALOG_TABS } from './components/ShopCatalogue';
+export { default as ShopSavingsModal } from './components/ShopSavingsModal';
+export * from './utils/effects';
